Added a new type of hook that can take arguments to trigger_hooks().
Added support for handling calls with the 'failed' extension. These calls can now be h
andled in the Adhearsion::Hooks::OnFailedCall hook.

Added a property to Call for checking if a call has failed. This is used by the DialPl
an::Manager

Made passing an entry point to the Call constructor optional.

Changed the name of the Builder class used to pull out context names from a dial plan.
 Its new name is ContextNameCollector

When creating a new context in a dialplan that has arguments given to its method, Kernel#method_missing will be called.

Refactored some of the stupid tests for framework hooks.

Added tests for SyntaxError issues in the dialplan, though I still haven't found what's eating the SyntaxErrors when a dialplan is syntactically invalid.


git-svn-id: http://svn.adhearsion.com/trunk@696 21f00dfd-7224-0410-ba9b-b8926ece8ce5