Greatly extended the available arguments to the input() dialplan method.
Here's the documentation for it:

This method is used to receive keypad input from the user. Digits are collected
via DTMF (keypad) input until one of three things happens:

 1. The number of digits you specify as the first argument is collected
 2. The timeout you specify with the :timeout option elapses.
 3. The "#" key (or the key you specify with :accept_key) is pressed

Usage examples

  input   # Receives digits until the caller presses the "#" key
  input 3 # Receives three digits. Can be 0-9, * or #
  input 5, :accept_key => "*"   # Receive at most 5 digits, stopping if '*' is pressed
  input 1, :timeout => 1.minute # Receive a single digit, returning an empty
                                  string if the timeout is encountered
  input 9, :timeout => 7, :accept_key => "0" # Receives nine digits, returning
                                             # when the timeout is encountered
                                             # or when the "0" key is pressed.

The :timeout option works like a digit timeout, therefore each digit pressed
causes the timer to reset. This is a much more user-friendly approach than an
absolute timeout.

Note that when you don't specify a digit limit, the :accept_key becomes "#"
because there'd be no other way to end the collection of digits. You can
obviously override this by passing in a new key with :accept_key.



git-svn-id: http://svn.adhearsion.com/trunk@717 21f00dfd-7224-0410-ba9b-b8926ece8ce5