Adhearsion::Calls (aka Adhearsion.active_calls) is now more Thread-safe and uses a Hash internally (instead of an Array). Also, it's more tested now.
An Adhearsion::Call now has a #unique_identifier() method. For Asterisk calls, this returns the "uniqueid" variable.

Call variable keys are now symbolized when a Call is instantiated

Each Call now has an @inbox property (implemented as a Queue). This will be more fleshed out soon. It will soon allow calls to communicate between each other in a Thread-safe fashion.


git-svn-id: http://svn.adhearsion.com/trunk@754 21f00dfd-7224-0410-ba9b-b8926ece8ce5