Adding the ability to specify gem versions in .ahnrc.
Did some refactoring of the Initializer. I've removed the initialization logic from the constructor, moved it to the start() instance method, and created a start() class method as sugar for new(*args).start.

Fixed a circumstance in which the last "ensure" closure of the AGI server raised a confusing exception if an exception was encountered in a BeforeCall hook.