Adjust all data buffer pointers when we slice the array.
This is the result of discussion in Lighthouse bug
#72.

No unit tests were broken in the implementation of
this fix.  The example code in the ticket now
functions correctly over several iterations.

One thing that may still need to be adjusted is
the size of the buffer.  The larger the buffer,
the slower AMI performance is.  This is due to the
many places string offsets are used in the parsing.

See: https://adhearsion.lighthouseapp.com/projects/5871-adhearsion/tickets/72-ami-lexer-buffer-offset#ticket-72-26