[FEATURE] Generate applications with an explicit dependency on Punchblock
* Many users of Adhearsion make direct use of Punchblock APIs (mostly command models) which are subject to Punchblock's own versioning.

* Adhearsion occasionally bumps the dependent Punchblock major version number in minor releases to gain access to new functionality that required such a bump in Punchblock, on the basis that this does not impact Adhearsion's API itself.

These two facts combine to produce the potential for breakage of an Adhearsion application by updating the Adhearsion minor version.

Generating an application with a specific dependency on Punchblock prevents bumping the Asterisk version if it requires a major version of Punchblock greater than the app was generated in, forcing the user to manually update the Punchblock version in their Gemfile at this point. This forces the user to consider their Punchblock API dependency and prepare for potential breakage.