Allow exceptions to escape CallController#execute
The goal is to allow spec runs to invoke controllers and get an exception when it occurs.  The current behavior catches the exception and sends it through the exception handler, which often will silently swallow it, especially if the tests complete and the process exits before the exception is worked from the queue.

This commit still has two broken tests, so more discussion is necessary.