Make latch waits consistent and configurable
I had noticed that several of the Travis CI test runs that failed recently had done so incorrectly because a `latch.wait` returned false instead of true. I expect this is because the Travis build-bots are very loaded and so the latches don’t clear fast enough. This makes the default close to what the specs already were, but gives Travis a little extra time to get the specs done. I’m hopeful this will cut down on those annoying false-positives.

Note that I only did the latches expected to return true because I didn’t want any latches that were *supposed* to stay latched to time out for other reasons.