Get rid of autoload in favor of explicit requires
In past discussion we’ve come to the conclusion that autoloading isn’t adding anything to the code since we already explicitly tell the autoloader the dependencies to load. In addition, in many cases we have to eager-load. All that autoload is doing is slowing down the normal process of requiring files.

The immediate cause for this work was the fact that `require ‘adhearsion/rspec’` was failing because of autoloader failing to resolve a dependency between `Adhearsion::CallController` and `Adhearsion::CallController::Dial`.

Along the way this also fixed some other previously masked bugs:

* PunchblockPlugin was defined as inheriting from Plugin in one place and not another
* At least three classes never existed and were never referenced, but were made available via autoload:
- Adhearsion::CallController::MenuDsl::Exceptions
- Adhearsion::Dispatcher
- Adhearsion::Plugin::Configuration