[CS] Upgrade RSpec on the way to 3.x
Something in the switch to expect() syntax has caused #should not to be defined on all objects, leaving this to be called as a method on the actor. Once these call actors are dead, however, they can't respond to this. We need to explicitly look at #alive?

We should review this tests suite's use of #should and indeed magic be_* matchers according to current RSpec best-practices.