Include simple rack-based HTTP server from Virginia
For tests we currently remove the generated Gemfile so that the app boots with the version of Adhearsion under test. The generated Gemfile contains sinatra such that the default rack app will work. In order for the app to boot with the sample Rack config, we must include Sinatra at the project Gemfile level.