Only log the new controller class name
Reference #563

These messages are printed at INFO level, and so should be more succinct. What we really care about is which controller is being invoked. Since the controller instances often include a large amount of metadata, these lines are often responsible for an unreasonable amount of log noise (due to the `#inspect`).