Don't trigger the exception event handler twice per exception
This resulted in exceptions being logged twice. The trigger in
CallController#execute! is not needed as this is already covered by
catching_standard_errors in CallController#bg_exec.

There was a test for this so I have adjusted it to check for the
on_error callbacks being triggered instead.

Fixes #618.