[CS] Stop passing now extraneous :async parameter to Client#execute_command.
The :async option to #execute_command was removed back in
adhearsion/punchblock@2b6c3ef
`[REMOVAL] Make Client#execute_command always async`.

Now, the #execute_command method is _always_ async.  Passing it in or not has no
change in behavior, but its use infers that there is an option for a synchronous
mode, which of course there no longer is.